# api-specs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Fruition
  description: Fruition PoC 백엔드 API
  title: Fruition API
  version: 0.0.1
security:
- bearerAuth: []
...
paths:
...
  /api/workspaces/{workspace_id}/assets/{asset_id}/content:
    get:
      description: 워크스페이스 멤버에게 관리 이미지 bytes를 반환합니다.
      operationId: getContent
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: asset_id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: 이미지 반환
        "304":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: 캐시된 이미지 사용
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: asset 또는 워크스페이스를 찾을 수 없음
      summary: 문서 이미지 조회
      tags:
      - Document Assets
...
  /api/workspaces/{workspace_id}/documents:
    get:
      description: 활성 문서의 호환용 평면 목록을 반환하며 파일명 검색을 지원합니다.
      operationId: list
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: query
        name: query
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
          description: 목록 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 서버 내부 오류
      summary: 문서 목록 조회
      tags:
      - Documents
    post:
      description: "PDF 또는 Markdown 파일을 업로드합니다. Markdown은 편집 상태와 처리 큐를 생성하고, PDF는 읽기 전용 원본으로만 저장합니다."
      operationId: upload
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 요청 멱등 키
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      - in: query
        name: folder_id
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
          description: 업로드 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 파일 없음 또는 잘못된 요청
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Idempotency-Key 충돌
        "415":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 지원하지 않는 파일 형식
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 서버 내부 오류
      summary: 문서 업로드
      tags:
      - Documents
...
  /api/workspaces/{workspace_id}/documents/{document_id}/content:
    put:
      description: 전체 Markdown과 신규 이미지를 저장합니다. base_revision이 현재 편집 revision과 일치할 때만 반영하며 revision_write_id 재시도는 기존 결과를 반환합니다. 이미지 포함 저장은 metadata part를 사용합니다.
      operationId: saveContent
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: document_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                apply_operation_id:
                  type: string
                  description: AI 편집 적용 표. `POST /agent/turn` 응답의 apply_operation_id를 그대로 전달하면 AI 작업 로그가 남습니다.
                base_revision:
                  type: string
                  description: 클라이언트가 조회한 현재 편집 revision
                markdown:
                  type: string
                  description: 저장할 전체 Markdown 본문
                metadata:
                  type: string
                  description: 이미지 포함 저장 metadata JSON
                revision_write_id:
                  type: string
                  description: 같은 저장 재시도에 재사용하는 write ID
                source:
                  type: string
                  description: "저장 출처. AI 편집 승인 시 \"agent\", 수동 저장 시 생략합니다."
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentContentSaveResponse"
          description: 저장 성공 또는 동일 본문 no-op
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "잘못된 Markdown, base_revision 또는 revision_write_id"
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 문서 소유자가 아님
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 문서 또는 워크스페이스를 찾을 수 없음
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 편집 revision 또는 revision_write_id 충돌
        "413":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Markdown 5MB 또는 이미지 제한 초과
        "415":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 지원하지 않는 이미지 형식
      summary: Markdown 본문 저장
      tags:
      - Documents
...
  /api/workspaces/{workspace_id}/documents/{document_id}/edit-lock:
    delete:
      description: 편집기 종료 시 호출한다. 보유자 본인의 잠금만 해제하며 멱등이다.
      operationId: release
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: document_id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: 잠금 해제 성공
      summary: 편집 잠금 해제
      tags:
      - Document Edit Lock
    post:
      description: 편집기 진입 시 호출한다. 비었거나 만료됐거나 본인 보유면 잠금을 부여(200)한다. 다른 사용자가 편집 중이면 423과 보유자 정보를 반환한다.
      operationId: acquire
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: document_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EditLockResponse"
          description: 잠금 획득/갱신
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 문서 소유자가 아님
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 문서 또는 워크스페이스를 찾을 수 없음
        "423":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EditLockResponse"
          description: 다른 사용자가 편집 중
      summary: 편집 잠금 획득
      tags:
      - Document Edit Lock
  /api/workspaces/{workspace_id}/documents/{document_id}/edit-lock/heartbeat:
    post:
      description: 편집 중 주기적으로 호출해 잠금을 연장한다. 보유자가 아니거나 만료됐으면 409.
      operationId: heartbeat
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - in: path
        name: document_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EditLockResponse"
          description: 잠금 연장
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 잠금 상실(만료/타인 보유)
      summary: 편집 잠금 heartbeat
      tags:
      - Document Edit Lock
...
  /api/workspaces/{workspace_id}/documents/{document_id}/position:
    patch:
      description: 문서를 대상 폴더와 정렬 위치로 이동합니다. base version과 Idempotency-Key로 동시 변경을 검증합니다.
      operationId: move_1
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - description: 이동할 문서 ID
        in: path
        name: document_id
        required: true
        schema:
          type: string
      - description: 요청 멱등 키
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentPositionRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentPositionResponse"
          description: 이동 성공 또는 멱등 재요청
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "잘못된 위치 또는 version, 또는 INVALID_IDEMPOTENCY_KEY(멱등 키 누락/유효하지 않음)"
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "문서, 대상 폴더 또는 워크스페이스를 찾을 수 없음"
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "version 충돌, IDEMPOTENCY_CONFLICT(동일 키에 다른 payload 사용) 또는 IDEMPOTENCY_IN_PROGRESS(활성 lease 재사용)"
      summary: 문서 위치 이동
      tags:
      - Documents
...
components:
  schemas:
...
    DocumentAttachmentSaveResponse:
      type: object
      description: 본문 저장과 함께 등록된 이미지 asset
      properties:
        asset_id:
          type: string
          format: uuid
          description: asset ID. 내려받기 경로에 넣는다.
        attachment_id:
          type: string
          format: uuid
          description: 문서-asset 연결 ID
        content_path:
          type: string
          description: Markdown 본문에서 이 이미지를 가리키는 경로
...
    DocumentContentSaveResponse:
      type: object
      properties:
        attachments:
          type: array
          description: 본문과 함께 저장된 이미지 asset 목록
          items:
            $ref: "#/components/schemas/DocumentAttachmentSaveResponse"
        changed:
          type: boolean
          description: 실제로 본문이 바뀌었는지 여부. 같은 내용이면 false다.
          example: true
        content_hash:
          type: string
          description: 저장된 본문의 해시. 같은 내용을 다시 저장했는지 판별에 쓴다.
        current_version:
          type: integer
          format: int64
          description: 저장 후 버전. 다음 저장의 base_version으로 쓴다.
          example: 4
        document_id:
          type: string
          description: 저장된 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        markdown:
          type: string
          description: 서버가 정규화한 뒤의 Markdown 본문
        updated_at:
          type: string
          format: date-time
          description: 저장 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
...
    DocumentItem:
      type: object
      description: 값이 없는 필드는 키 자체가 빠진다. null 체크가 아니라 존재 여부로 확인한다.
      properties:
        area:
          type: string
          description: 문서가 속한 영역
        byte_size:
          type: integer
          format: int64
          description: 파일 크기(바이트)
          example: 482913
        current_version:
          type: integer
          format: int64
          description: 낙관적 잠금 버전. 이후 쓰기 요청의 base_version에 그대로 넣는다.
          example: 1
        display_name:
          type: string
          description: 화면에 보여줄 이름
          example: 설계문서
        document_role:
          type: string
          description: "문서 역할. Markdown은 EDITABLE, PDF는 ORIGINAL이다."
          enum:
          - EDITABLE
          - ORIGINAL
        editable:
          type: boolean
          description: 본문 편집이 가능한지 여부
          example: false
        error_message:
          type: string
          description: status=failed일 때의 실패 사유
        extracted_text_uri:
          type: string
          description: 추출된 본문 텍스트가 저장된 경로. 처리 전이면 없다.
        file_type:
          type: string
          description: 확장자 기준 파일 종류
          example: pdf
        filename:
          type: string
          description: 저장된 파일명
          example: 설계문서.pdf
        id:
          type: string
          description: 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        item_kind:
          type: string
          description: 트리에서의 항목 종류
          example: document
        mime_type:
          type: string
          description: 파일 MIME 타입
          example: application/pdf
        needs_reingest:
          type: boolean
          description: 마지막 ingest 이후 편집본이 바뀌어 재분석이 필요한지 여부. 편집 가능 문서만 true가 될 수 있다.
          example: false
        pipeline_run_id:
          type: string
          description: 이 문서를 처리한 pipeline run ID
        processed_at:
          type: string
          format: date-time
          description: 처리 완료 시각(ISO-8601 UTC). 처리 전이면 없다.
        processing_stage:
          type: string
          description: 현재 처리 단계 이름
          example: ingest
        processing_started_at:
          type: string
          format: date-time
          description: 현재 처리 작업의 시작 시각(ISO-8601 UTC). 처리 시작 전이면 없다.
        processing_state:
          type: string
          description: 처리 진행 상태. status보다 세분화된 값이다.
          enum:
          - starting
          - running
          - stalled
          - completed
          - failed
        source_document_id:
          type: string
          description: 변환으로 만들어진 문서라면 원본 문서 ID
        source_uri:
          type: string
          description: 원본 파일이 저장된 오브젝트 스토리지 경로
        status:
          type: string
          description: 문서 처리 상태
          enum:
          - uploaded
          - processing
          - completed
          - failed
        updated_at:
          type: string
          format: date-time
          description: 마지막 변경 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
        uploaded_at:
          type: string
          format: date-time
          description: 업로드 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
...
    DocumentListResponse:
      type: object
      properties:
        documents:
          type: array
          description: 활성 문서 목록. 휴지통 문서는 포함되지 않는다.
          items:
            $ref: "#/components/schemas/DocumentItem"
    DocumentPositionRequest:
      type: object
      description: 문서 이동·순서 변경 요청
      properties:
        base_version:
          type: integer
          format: int64
          description: 직전에 읽은 current_version. 서버 값과 다르면 409다.
          example: 1
        folder_id:
          type: string
          format: uuid
          description: 옮길 폴더 ID. null이면 루트로 옮긴다.
          example: 8d4f1e6c-3b0a-497d-25e4-f831b9f4c7e2
        position:
          type: integer
          format: int32
          description: 새 폴더 안에서의 0-based 삽입 위치. 생략하면 맨 뒤에 놓는다.
          example: 0
      required:
      - base_version
    DocumentPositionResponse:
      type: object
      properties:
        current_version:
          type: integer
          format: int64
          description: 이동 후 버전
          example: 2
        folder_id:
          type: string
          format: uuid
          description: 이동 후 폴더 ID. 루트면 null이다.
        id:
          type: string
          description: 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        sort_order:
          type: integer
          format: int64
          description: 이동 후 정렬 순서
          example: 1024
...
    DocumentUploadResponse:
      type: object
      properties:
        byte_size:
          type: integer
          format: int64
          description: 파일 크기(바이트)
          example: 482913
        current_version:
          type: integer
          format: int64
          description: 낙관적 잠금 버전. 이후 쓰기 요청의 base_version에 그대로 넣는다.
          example: 1
        document_role:
          type: string
          description: "문서 역할. Markdown은 EDITABLE, PDF는 ORIGINAL이다."
          enum:
          - EDITABLE
          - ORIGINAL
        editable:
          type: boolean
          description: "본문 편집이 가능한지 여부. Markdown은 true, PDF는 false다."
          example: false
        filename:
          type: string
          description: 업로드된 원본 파일명
          example: 설계문서.pdf
        id:
          type: string
          description: 문서 ID
          example: doc_1b9f4c7e2a8d4f1e6c3b0a97d25e4f83
        mime_type:
          type: string
          description: 파일 MIME 타입. PDF와 Markdown만 허용되며 그 외는 415다.
          example: application/pdf
        source_uri:
          type: string
          description: 원본 파일이 저장된 오브젝트 스토리지 경로
        status:
          type: string
          description: 문서 처리 상태
          enum:
          - uploaded
          - processing
          - completed
          - failed
        uploaded_at:
          type: string
          format: date-time
          description: 업로드 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
...
    EditLockResponse:
      type: object
      description: 편집 잠금 상태. 잠금이 없으면 필드 자체가 빠진다.
      properties:
        expires_at:
          type: string
          format: date-time
          description: 잠금 만료 시각(ISO-8601 UTC). 이 시각이 지나면 다른 사용자가 잠글 수 있다.
          example: 2026-08-13T04:25:24.371948Z
        holder_display_name:
          type: string
          description: 현재 편집 중인 사용자의 표시 이름
          example: 표시 이름
        holder_user_id:
          type: string
          description: 현재 편집 중인 사용자 ID
          example: user_3f1c8a6b52d7411e9c04ab5d2e7f6081
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 분기 기준으로 쓰는 값. message가 아니라 이 값으로 분기한다.
          example: INVALID_REQUEST
        details:
          type: array
          description: 검증 실패(400 INVALID_REQUEST)일 때만 존재한다. 값이 없으면 키 자체가 빠지므로 null 체크가 아니라 존재 여부로 확인한다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: 사용자에게 그대로 보여도 되는 한국어 문구
          example: 요청 형식이 올바르지 않습니다.
    ErrorResponse:
      type: object
      description: 컨트롤러까지 도달한 요청이 실패했을 때 모든 API가 공유하는 에러 응답 형식. 인증 필터가 막는 401은 HttpStatusEntryPoint가 상태 코드만 내보내므로 이 본문이 없다.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 에러 상세
...
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: 문제가 된 요청 필드명
          example: email
        reason:
          type: string
          description: 그 필드의 위반 사유
          example: email은 필수입니다.
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http